pyproject.toml
[tool.poetry]
name = "puppyplace-products-api"
version = "0.1.0"
description = ""
authors = ["Beatriz Uezu <beatriz.uezu@gmail.com>"]

[tool.isort]
line_length = 110
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_local_folder = "apps,tests"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"

[tool.poetry.dependencies]
python = "^3.9"

dj-database-url = "^0.5.0"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
mssql-django = "^1.0"
prettyconf = "^2.2.1"
pyodbc = "^4.0.32"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
blue = "^0.7.0"
factory-boy = "^3.2.0"
ipdb = "^0.13.9"
ipython = "^7.27.0"
isort = "^5.9.3"
pre-commit = "^2.15.0"
pytest = "^5.2"
pytest-cov = "^2.12.1"
pytest-deadfixtures = "^2.2.1"
pytest-django = "^4.4.0"
pytest-icdiff = "^0.5"
pytest-envvars = "^1.2.1"
pytest-randomly = "^3.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
